Fix issue with node 10 and options.ca issue
The json.stringify json.parse object clone methods changes the type of
the ca option in a way that the http module of node 10 does not support.
Since this is just a ugly fork and only used in one single place,
(ikano integration), simply remove the clone. This will of course make
this code modify the passed in options object but I can live with that.

AB-1351